+++ src/main/java/insane96mcp/enhancedai/setup/EAAttributes.java
@@ -10,5 +10,6 @@ import net.minecraftforge.registries.RegistryObject;
 public class EAAttributes {
 	public static final DeferredRegister<Attribute> ATTRIBUTES = DeferredRegister.create(ForgeRegistries.ATTRIBUTES, EnhancedAI.MOD_ID);
 
+	//TODO remove generic.
 	public static final RegistryObject<Attribute> XRAY_FOLLOW_RANGE = ATTRIBUTES.register("generic.xray_follow_range", () -> new RangedAttribute("attribute.name.generic.xray_follow_range", 0d, 0d, 256d));
 }