+++ changelog.md
@@ -2,7 +2,13 @@
 
 ## Upcoming
 * Added a new explosion sound for Angry Creepers. It's now the default
-  * The fuse sound is vanilla but the explosion sound is the old minecraft beta one
+    * The fuse sound is vanilla but the explosion sound is the old minecraft beta one
+
+## 2.4.10
+* Removed `use_follow_range_changes` entity type tag
+  * Replaced with `change_follow_range` that applies only the follow range override
+  * Also with `apply_xray` that applies only the xray range override
+* Added "Block Blacklist as Whitelist" for Miner Mobs
 
 ## 2.4.9
 * Fixed Beta creepers ignoring movement speed changes